Reference full domain name for upstream projects
Because of a bug in tomcat 8.5.31, it will always respond
with a status 400 to requests with a - in the last part of
the hostname. As we are referencing the upstream servers with
<servicename>.<workspacename> by default, and the workspace name
contains a dash on ci and demo, upstream requests fail.

For that reason we now reference the full domain name. See also
https://stackoverflow.com/questions/51023105/spring-boot-service-in-kubernetes-always-responses-with-http-status-400