Allow access to services using an Authorization header for api clients
The default mechanism to check for user credentials is using a session
cookie. If a user is logged in, the cookie is being checked. However,
the system should also allow access if a proper JWT token is provided
in the Authorization header. This is apparently non-trivial. See
https://stackoverflow.com/questions/34718586/spring-security-oauth2-enableoauth2sso-but-accept-tokens-as-authentication-t

This fix is required for VRE-65