Allow access_token to be refreshed
The access_token will be refreshed when a call to an upstream
server is being made with an expired token. However, the refresh
call is only being made if the client_id stored with the original
request is the same as the client_id from the configuration.
This code effectively adds the client_id that was used to retrieve
the token into the stored request.