Rewrite destination header to allow moving and copying files
The webdav-server implementation works with files with a certain root
(/api/storage/webdav). Its express.js middleware also removes that part
from the request url, so the webdav server does not have to care about it.
However, that part is still present in the Destination header, when moving
or copying files. For that reason, it should be removed as well.