Stop replacing tag and repository in values.yaml
The replacement is not needed in this helm chart, as it is normally
used for matching the helm chart version with the docker image version.
As this chart does not have its own docker images, there is no need to
do the replacement. Moreover, it replaced tags for other images
(e.g. JupyterHub) as well, which caused errors.