Update jsonld dependency
The previous version of the jsonld library has a hard dependency on
native code. This code is never used in the browser, but building it
required some additional steps and time.

The new version of the library has an optional dependency on the
native code, and should be able to build without additional effort.

See also https://github.com/digitalbazaar/rdf-canonize/issues/5