Remove validation on refresh tokens
As explained in the keycloak documentation at
https://www.keycloak.org/docs/4.6/server_admin/index.html#realm_keys
there is only one key active and returned at a time. This means that our
application can only validate tokens with one key.

As there is also no real need to validate the refresh token (it is done
by keycloak anyway), the validation is removed from this library.