Move logic for file operations into VFS resource objects
The logic for creation of new files and directories and for
updating the file contents was exposed in the public API of
the resourceFactory. Although some of the logic is still in
the resourceFactory to keep related logic together, the public
API for doing these operations has moved to the VfsResource
objects. This makes it easier to interact with the resource objects