Return only workspace users from pluto endpoint
The call to keycloak previously returned all keycloak users. We would
want to retrieve all users with a specific role, but keycloak does not
provide such an endpoint atm.

However, our keycloak setup is such that all users for a workspace are
member of a certain group. Keycloak does provide an endpoint to return
all members of a group, which is now used. To retrieve those members,
we do need the group id, which is not known on installation. For that reason
the groupId is retrieved from keycloak on every first call to this endpoint
and cached forever.

Please note that the mockito upgrade is needed to prevent errors on JDK 11.

Fixes VRE-446