Add new SHACL vocabularies to saturn
The format of the vocabularies has changed according to the SHACL
standards. See VRE-617 and VRE-624. The updated vocabularies have
been added to saturn

The vocabulary has been split into a fixed part (system-vocabulary)
and a part that the user can change (user-vocabulary). Both are exposed
in separate endpoints, where only the user-vocabulary can be updated.