Detach new entity dialog from dropdown component
By detaching the dropdown from the dialog, the system is capable
of handling multiple dialogs of the same type at once. Previously
the container would be rendered again if a new entity of that type
was added. However, it would then also rerender the dialog. If that
dialog would also have an open dialog, it would disappear.

This solution ensures that the new entity dialog will not be rerendered
if only the dropdown changes.