Change sh:class property to generic IRI
We currently do not have any Class entities in our vocabulary,
as we only work with shapes. However, this disables us to show
a dropdown of Class entities to choose from. For that reason, the
sh:class is now represented as generic IRI, where the user can
enter an arbitrary URI.

In the future this is likely to change, as we add more specific
functionality to the vocabulary editor.