Only show table values for metadata when it is an Array
If something went wrong with parsing or normalizing the jsonLD object,
it can happen that we try to render an object in a table cell. As this
renders as [object Object], this is prevented.

One situation where this occurs is when there is an empty string value
for one of the fields in the table.