Allow unauthorized requests to frontend urls
Our frontend is accessible for unauthorized users that are logged in
(i.e. users that have an account in keycloak, but are not allowed to
access the workspace). They see a proper error message in the frontend.

Our frontend is also served if a user accesses a frontend url directly
(e.g. /metadata or /collections/c1). This ensures that the user is able
to refresh the page or share a link, and the frontend is still working.

This fix ensures that these urls are also accessible if the user is not
really authorized to see the content. The frontend will show the correct
error message.

Fixes VRE-436