Limit the number of metadata triples being returned.
For performance reasons, we do not want regular user to be able to fetch
the full database. This change limits the number of triples being returned.

Please note that in some cases the number of triples being returned
is larger than this limit due to the way the query is setup. For example,
when object properties are asked for, this limit restricts the number of
triples being retrieved from the database, but not taking into account
those object properties.

Fixes VRE-891