Remove Jupyter from workspace.
JupyterHub has been moved to a separate repository, with its own scripts.
This separates the development of the jupyter integration from the workspace,
as it has a different lifecycle. This also allows us to install jupyter in a
separate k8s namespace or cluster, with additional security measures.

Fixes VRE-496