Add custom hook to keep track of isMounted state
The custom hook makes it simple to check whether a component is still mounted.
In many cases, we update the state of a component in a promise callback. Due to
the asynchronous nature of promises, we are never certain that the component is still
mounted when the callback is executed. React warns about changing the state of
a component that is unmounted. This hook allows for an easy check whether the
component is unmounted, without polluting the component itself.