Add usernames to permissions on every render.
The usernames are added to the permissions based on the retrieved usernames.
Previously, this was only done if the call to fetch permissions finished. However,
any updates to the list of users after that was not taken into account. This
change makes sure that changes in the list of users are reflected in the permissions
as well. Fixes VRE-1046