Always allow embedding of important properties of linked entities.
This allows users without public metadata access or access to a
particular workspace to see the label (for workspace also description)
of entities linked to an entity that the user has access to.
E.g., when selecting a collection, the user can always see the label
of the owner workspace and name of creating or modifying user, and the
label of other linked entities (e.g., analysis type, sample, subject).

Fixes VRE-1491.